SP0 final-review fixes: u64 readback arithmetic + dimension guard + doc/tidy
Whole-branch review (verdict: ship with minor follow-ups) surfaced:
- Important: readback size was computed as u32 (width*height*4) before the u64
  cast, so untrusted client dimensions could wrap and undersize the buffer.
  Now widened per-factor: correct by construction. Added a MAX_DIMENSION guard
  in handle_connection so an oversized BeginFrame errors cleanly before the GPU,
  plus a regression test (absurdly_large_dimensions_are_rejected_before_the_gpu).
- Minor: corrected e2e.rs comment/doc (server returns on EndFrame, not EOF; the
  write shutdown is hygiene). Dropped redundant rayland-wire dev-dependency.

Deferred (logged, not blocking): lib exposing anyhow::Result (thiserror once
render becomes a reusable API in SP1+); optional HOST_READ barrier for
sync-validation; VK_ICD_FILENAMES -> VK_DRIVER_FILES in CI.

+++ crates/rayland-server/src/render.rs
@@ -350,7 +350,11 @@ unsafe fn render_triangle_inner(request: &FrameRequest) -> anyhow::Result<Render
     unsafe { device.unmap_memory(vbuf_mem) };
 
     // --- Readback buffer (host-visible, holds the tightly-packed image after the copy) ---
-    let readback_size = (request.width * request.height * 4) as u64;
+    // Compute the size in u64 arithmetic. width/height arrive from an untrusted client
+    // BeginFrame; multiplying them as u32 first could wrap (e.g. 46341*46341*4), silently
+    // sizing the buffer too small in release builds. Widening each factor before the
+    // multiply makes the arithmetic correct by construction regardless of the inputs.
+    let readback_size = request.width as u64 * request.height as u64 * 4;
     let rbuf_info = vk::BufferCreateInfo::default()
         .size(readback_size)
         .usage(vk::BufferUsageFlags::TRANSFER_DST)
